requirements.txt
streamlit
langchain-community
faiss-cpu
sentence-transformers
openai
datasets
python-dotenv